+++ changelog.md
@@ -1,4 +1,7 @@
 # Fake Rate Changelog
+## v1.1.5 (2024-05-25)
+- Fixed incompatibility with the mod "Demons In Between" by me
+
 ## v1.1.4 (2024-05-24)
 - Fixed incompatibility with the mod "Underrated Non-Demons" by acaruso
 - Tweaked appearance of the fake rate button